Migrate long-running test to new cluster to enable monitoring (#7357)

+++ .github/scripts/cleanup-cluster.sh
@@ -21,22 +21,26 @@ set -e
 echo "cleaning up cluster"
 
 # Delete all test resources in queuemessages.
-echo "delete all resources in queuemessages.ucp.dev"
-kubectl delete queuemessages.ucp.dev -n radius-system --all
+if kubectl get crd queuemessages.ucp.dev > /dev/null 2>&1; then
+    echo "delete all resources in queuemessages.ucp.dev"
+    kubectl delete queuemessages.ucp.dev -n radius-system --all
+fi
 
 # Testing deletion of deployment.apps.
 
 # Delete all test resources in resources without proxy resource.
-echo "delete all resources in resources.ucp.dev"
-resources=$(kubectl get resources.ucp.dev -n radius-system --no-headers -o custom-columns=":metadata.name")
-for r in $resources; do
-    if [[ $r == scope.local.* || $r == scope.aws.* || -z "$r" ]]; then
-        echo "skip deletion: $r"
-    else
-        echo "delete resource: $r"
-        kubectl delete resources.ucp.dev $r -n radius-system --ignore-not-found=true
-    fi
-done
+if kubectl get crd resources.ucp.dev > /dev/null 2>&1; then
+    echo "delete all resources in resources.ucp.dev"
+    resources=$(kubectl get resources.ucp.dev -n radius-system --no-headers -o custom-columns=":metadata.name")
+    for r in $resources; do
+        if [[ $r == scope.local.* || $r == scope.aws.* || -z "$r" ]]; then
+            echo "skip deletion: $r"
+        else
+            echo "delete resource: $r"
+            kubectl delete resources.ucp.dev $r -n radius-system --ignore-not-found=true
+        fi
+    done
+fi
 
 # Delete all test namespaces.
 echo "delete all test namespaces"
